+++ backend/src/usr/peru_short/peru.ts
@@ -6,11 +6,28 @@ import { step as eqDmg } from './5_eqdamage/eqdamage';
 import { step as tsSim } from './6_tssim/tssim';
 import { step as tsDmg } from './7_tsdamage/tsdamage';
 import { step as sysRel } from './8_sysrel/sysrel';
+import config from "../../config.json";
+import axios from "axios";
+
 
 export const peruShortFactory = new ScenarioFactory(
     'PeruShort',
     'Peru Scenario Description',
 );
+
+peruShortFactory.registerCondition(async () => {
+    for (const [key, val] of Object.entries(config.services)) {
+        const url = val.url;
+        const procId = val.id;
+        const request = `${url}?service=WPS&request=DescribeProcess&identifier=${procId}&version=1.0.0`;
+        try {
+            const response = await axios.get(request);
+        } catch (error) {
+            return `Error in attempting to access ${url} for process ${key} (id=${procId}): ${error}`;
+        }
+    }
+    return true;
+});
 
 peruShortFactory.registerStep(select);
 peruShortFactory.registerStep(eqSim);